% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assayBoxPlot}
\alias{assayBoxPlot}
\title{Boxplot of Counts}
\usage{
assayBoxPlot(
  TapestriExperiment,
  alt.exp = NULL,
  assay = NULL,
  log.y = T,
  split.features = F,
  coldata.set = NULL
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{alt.exp}{Chr string indicating altExp to use, NULL (default) uses main experiment.}

\item{assay}{Chr string indicating assay to use. NULL (default) selects first listed assay.}

\item{log.y}{Logical, if TRUE, scales data by \code{log1p()}. Default TRUE.}

\item{split.features}{Logical, if TRUE, splits plot by rowData features. Default FALSE.}

\item{coldata.set}{Chr string indicating colData column to use for X axis categories. Default NULL.}
}
\value{
ggplot object
}
\description{
Draws boxplot of matrix data. Useful for visualizing altExp count data.
}
\examples{
\dontrun{assayBoxPlot(TapestriExperiment, "chrYCounts", assay = "counts",
split.features = T, coldata.set = "cluster")}
}
