% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtureCalls.R
\name{calcGMMCopyNumber}
\alias{calcGMMCopyNumber}
\title{GMM}
\usage{
calcGMMCopyNumber(
  TapestriExperiment,
  cell.barcodes,
  control.copy.number,
  model.components = 1:5,
  model.priors = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
GMM
}
