% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumberCalculation.R
\name{calcSmoothCopyNumber}
\alias{calcSmoothCopyNumber}
\title{Smooth copy numbers values by chromosome and chromosome arm}
\usage{
calcSmoothCopyNumber(TapestriExperiment, method = "median")
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object.}

\item{method}{Character, smoothing method. Supports median (default) or mean.}
}
\value{
\code{TapestriExperiment} with \code{smoothedCopyNumber} and \code{discreteCopyNumber} assays in \code{altExp} slots \code{smoothedCopyNumberByChr} and \code{smoothedCopyNumberByArm}.
}
\description{
\code{calcSmoothCopyNumber()} takes \code{copyNumber} slot values for probes on a chromosome and smoothes them by median (default) for each chromosome
and chromosome arm, resulting in one copy number value per chromosome and chromosome arm for each cell barcode.
Values are then discretized into integers by conventional rounding.
Smoothed copy number and discretized smoothed copy number values are stored as \code{smoothedCopyNumber} and \code{discreteCopyNumber} assays,
in \code{altExp} slots \code{smoothedCopyNumberByChr} for chromosome-level smoothing, and \code{smoothedCopyNumberByArm} for chromosome arm-level smoothing.
}
\examples{
\dontrun{
TapestriExperiment <- calcSmoothCopyNumber(TapestriExperiment)
}
}
