% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNweaveR-package.R
\docType{package}
\name{CNweaveR-package}
\alias{CNweaveR}
\alias{CNweaveR-package}
\title{CNweaveR: DNA Copy Number Analysis for Genome-wide Tapestri Panels}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

CNweaveR enables analysis of DNA copy number variants (CNVs) in thousands of single cells, using custom genome-wide targeted DNA sequencing panels for the Mission Bio Tapestri system. This package allows users to easily parse, manipulate, and visualize datasets produced from the automated Tapestri Pipeline, with support for normalization, clustering, and ploidy calling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeymays/CNweaveR}
  \item \url{http://joeymays.xyz/CNweaveR/}
  \item Report bugs at \url{https://github.com/joeymays/CNweaveR/issues}
}

}
\author{
\strong{Maintainer}: Joseph Mays \email{josephcmays@gmail.com} (\href{https://orcid.org/0000-0003-4903-938X}{ORCID})

}
\keyword{internal}
