% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{corner}
\alias{corner}
\title{Glimpse the top-left corner of a matrix}
\usage{
corner(input.mat)
}
\arguments{
\item{input.mat}{a matrix-like object}
}
\value{
matrix-like object matching input class
}
\description{
Outputs up to 5 rows and columns of the input matrix object to get a quick look without filling the console.
}
\examples{
\dontrun{
corner(assay(TapestriObject))
}
}
