% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{countBoxPlot}
\alias{countBoxPlot}
\title{Boxplot of Counts}
\usage{
countBoxPlot(
  TapestriExperiment,
  alt.exp = NULL,
  assay = "counts",
  log.y = T,
  split.features = F,
  coldata.set = NULL
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{alt.exp}{Chr string indicating altExp to use, NULL (default) uses main experiment.}

\item{assay}{Chr string indicating assay to use. Default "counts".}

\item{log.y}{Logical, if TRUE, scales counts by \code{log1p()}. Default TRUE.}

\item{split.features}{Logical, if TRUE, splits plot by rowData features.}

\item{coldata.set}{Chr string indicating colData column to use for X axis categories. Default NULL.}
}
\value{
ggplot object
}
\description{
Draws boxplot of count data. Useful for visualizing altExp count data.
}
\examples{
\dontrun{countBoxPlot(TapestriExperiment, "chrYCounts", assay = "counts",
split.features = T, coldata.set = "cluster")}
}
