% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCytobands.R
\name{getCytobands}
\alias{getCytobands}
\title{Retrieve and add chromosome cytobands and chromosome arms to probe metadata of TapestriExperiment object}
\usage{
getCytobands(TapestriExperiment, genome = "hg19")
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{genome}{Character string indicating reference genome to use. Only hg19 is currently supported.}
}
\value{
TapestriExperiment object with rowData updated to include chromosome arms and cytobands
}
\description{
Retrieve and add chromosome cytobands and chromosome arms to probe metadata of TapestriExperiment object
}
\examples{
\dontrun{tapObject <- getCytobands(tapObject, genome = "hg19")}
}
