% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyCalcuation.R
\name{generateControlPloidyTemplate}
\alias{generateControlPloidyTemplate}
\alias{getPloidy}
\title{Generate Control Ploidy Template}
\usage{
generateControlPloidyTemplate(ploidy.all = 2, sample.label.all = "cluster")

getPloidy(TapestriExperiment, control.ploidy, coldata.set = "cluster")
}
\arguments{
\item{ploidy.all}{Numeric, sets all entries of \code{ploidy} in the output. Default 2.}

\item{sample.label.all}{Chr string, sets all entries of \code{sample.label} in the output. Default "cluster".}

\item{TapestriExperiment}{TapestriExperiment object.}

\item{control.ploidy}{A data.frame with columns \code{arm}, \code{ploidy}, and \code{sample.label}. See description.}

\item{coldata.set}{A chr string indicating colData column to use for subsetting cell.barcodes. Default "cluster".}
}
\value{
data.frame with 3 columns named \code{arm}, \code{ploidy}, and \code{sample.label}

TapestriExperiment object with ploidy values in \code{ploidy} assay slot.
}
\description{
\code{generateControlPloidyTemplate()} generates a data.frame template for use in \code{\link[=getPloidy]{getPloidy()}}.

\code{getPloidy()} transforms the normalized count matrix of a TapestriExperiment object into ploidy values based on a reference subset of cell barcodes and given ploidy value (e.g. 2 for diploid).
This is practically used to set the median ploidy of a control, usually diploid, cell population to a known ploidy value, e.g. 2, and then calculate the ploidy for all the cells relative to that control population.
This occurs individially for each probe.
\code{control.ploidy} is a lookup table (data.frame) to indicate the ploidy value and cell barcodes to use as the reference.
A template for \code{control.ploidy} can be generated using \code{\link[=generateControlPloidyTemplate]{generateControlPloidyTemplate()}}.
The \code{control.ploidy} data.frame should include 3 columns named \code{arm}, \code{ploidy}, and \code{sample.label}.
\code{arm} is chromosome arms from chr1p through chrXq, \code{ploidy} is the reference ploidy value, and \code{sample.label} is the value corresponding to the colData column given in \code{coldata.set} to indicate the cell barcode subset to use to set the ploidy.
}
\section{Functions}{
\itemize{
\item \code{generateControlPloidyTemplate()}: 

}}
\examples{
\dontrun{control.ploidy <- generateControlPloidyTemplate()}
\dontrun{TapestriExperiment <- getPloidy(TapestriExperiment,
control.ploidy, coldata.set = "cluster")}
}
