% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotapR-package.R
\docType{package}
\name{karyotapR-package}
\alias{karyotapR}
\alias{karyotapR-package}
\title{karyotapR: DNA Copy Number Analysis for Genome-Wide Tapestri Panels}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analysis of DNA copy number in thousands of single cells using custom genome-wide targeted DNA sequencing panels for the Mission Bio Tapestri platform. Users can easily parse, manipulate, and visualize datasets produced from the automated Tapestri Pipeline, with support for normalization, clustering, and copy number calling. Functions are also available to deconvolute multiplexed samples by genotype and parsing barcoded reads from exogenous lentiviral constructs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeymays/karyotapR}
  \item \url{http://joeymays.xyz/karyotapR/}
  \item Report bugs at \url{https://github.com/joeymays/karyotapR/issues}
}

}
\author{
\strong{Maintainer}: Joseph Mays \email{josephcmays@gmail.com} (\href{https://orcid.org/0000-0003-4903-938X}{ORCID})

}
\keyword{internal}
