% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTapestriExperiment.R
\name{newTapestriExperimentExample}
\alias{newTapestriExperimentExample}
\title{Create Example \code{TapestriExperiment}}
\usage{
newTapestriExperimentExample()
}
\value{
\code{TapestriExperiment} object.
}
\description{
Creates a \code{TapestriExperiment} object for demonstration purposes,
which includes 240 probes across the genome, and 300 cells of 3 types.
Raw counts are generated randomly.
Type 1 has 75 cells, all XY, all diploid.
Type 2 has 100 cells, all XX, with 3 copies of chr 7, otherwise diploid.
Type 3 has 125 cells, all XY, with 1 copy of chr 1p, otherwise diploid.
}
\examples{
tapExperiment <- newTapestriExperimentExample()
}
