% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBarcodedReads.R
\name{parseBarcodedReadsFromContig}
\alias{parseBarcodedReadsFromContig}
\alias{parseBarcodedReads}
\title{Parse Barcoded Reads}
\usage{
parseBarcodedReadsFromContig(
  bam.file,
  barcode.lookup,
  contig,
  cell.barcode.tag = "RG",
  max.mismatch = 2,
  with.indels = T
)

parseBarcodedReads(
  TapestriExperiment,
  bam.file,
  barcode.lookup,
  probe.tag,
  return.table = F,
  ...
)
}
\arguments{
\item{bam.file}{Chr string indicating file path of BAM file. \code{.bai} BAM index file must be in the same location.}

\item{barcode.lookup}{data.frame where the first column is the barcode identifier/name and the second column is the DNA sequence. Headers are ignored.}

\item{contig}{Chr string of contig or chromosome name to search for barcodes in. Can be a vector of more than one contig to expand search space.}

\item{cell.barcode.tag}{Character string of length 2, indicates cell barcode field in BAM, specified by Tapestri pipeline (currently "RG"). Default "RG".}

\item{max.mismatch}{The maximum and minimum number of mismatching letters allowed. See \code{\link[Biostrings:matchPattern]{Biostrings::matchPattern()}}.}

\item{with.indels}{If TRUE then indels are allowed. See \code{\link[Biostrings:matchPattern]{Biostrings::matchPattern()}}.}

\item{TapestriExperiment}{TapestriExperiment object}

\item{probe.tag}{Chr string, either "gRNA" or "sample.barcode" to parse counts from grnaCounts or sampleBarcodeCounts alternative experiments, respectively.}

\item{return.table}{If TRUE, returns table of read counts per barcode. If FALSE, returns TapestriExperiment. Default FALSE.}

\item{...}{Arguments to pass on to \code{parseBarcodedReadsFromContig()}.}
}
\value{
A data.frame of read counts for each specified barcode.

An updated TapestriExperiment object with read counts added to the colData slot. If \code{return.table == TRUE}, a data.frame of read counts for each specified barcode.
}
\description{
\code{parseBarcodedReads()} and \code{parseBarcodedReadsFromContig()} match exogenous DNA barcode sequences to their associated cell barcodes and saves them to the colData (cell barcode metadata) of the input TapestriExperiment object.
\code{parseBarcodedReads()} is a shortcut for \code{parseBarcodedReadsFromContig()}, allowing the user to specify 'gRNA' or 'sample.barcode', and assign the result to the input object..
The entries in the \code{barcode.lookup} table do not have to be present in the sample, allowing users to keep one master table/file of available barcode sequences to use for all experiments.
}
\examples{
\dontrun{counts <- parseBarcodedReadsFromContig(bam.file, barcode.lookup, "virus_ref2")}
\dontrun{counts <- parseBarcodedReads(TapestriExperiment,
bam.file, barcode.lookup, "gRNA")}
}
