% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtureCalls.R
\name{plotCopyNumberGMM}
\alias{plotCopyNumberGMM}
\title{Plot Copy Number GMM Components}
\usage{
plotCopyNumberGMM(
  TapestriExperiment,
  feature.id = 1,
  chromosome.scope = "chr",
  draw.boundaries = F
)
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object}

\item{chromosome.scope}{}
}
\description{
Plot Copy Number GMM Components
}
