% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reducedDimPlot}
\alias{reducedDimPlot}
\title{Scatter plot for dimensional reduction results.}
\usage{
reducedDimPlot(
  TapestriExperiment,
  alt.exp = "alleleFrequency",
  dim.reduction,
  dim.x = 1,
  dim.y = 2,
  group.label = NULL
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{alt.exp}{Chr string indicating altExp to use, NULL uses top-level experiment. Default "alleleFrequency".}

\item{dim.reduction}{Chr, which dimension reduction to plot, either "PCA" or "UMAP".}

\item{dim.x}{Numeric, index of dimensional reduction data to plot on X axis. Default 1.}

\item{dim.y}{Numeric, index of dimensional reduction data to plot on Y axis. Default 2.}

\item{group.label}{Chr string indicating colData column for coloring samples. Default NULL.}
}
\value{
ggplot scatter plot
}
\description{
Scatter plot for dimensional reduction results.
}
\examples{
\dontrun{reducedDimPlot(TapestriExperiment, dim.reduction = "pca")}
}
