% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{runPCA}
\alias{runPCA}
\title{Clustering by PCA}
\usage{
runPCA(
  TapestriExperiment,
  feature.set = "alleleFrequency",
  sd.min.threshold = 0,
  center = T,
  scale. = T
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{feature.set}{Chr string identifying the altExp feature set to perform PCA on. Default "alleleFrequency".}

\item{sd.min.threshold}{Numeric, minimum threshold for allelefreq.sd. Increase to run PCA on fewer, increasingly variable dimensions. Default 0.}

\item{center}{Logical, whether the variables should be shifted to be zero centered. See \code{\link[stats:prcomp]{stats::prcomp()}}.}

\item{scale.}{Logical, whether the variables should be scaled to have unit variance before the analysis takes place. See \code{\link[stats:prcomp]{stats::prcomp()}}.}
}
\value{
TapestriExperiment with PCA results saved to reducedDims slot of altExp, and proportion of variance saved to metadata slot of altExp.
}
\description{
Runs a set of features through PCA and saves results to reducedDims slot of input TapestriObject.
}
\examples{
\dontrun{TapestriExperiment <- runPCA(TapestriExperiment, sd.min.threshold = 35)}
}
