% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{runUMAP}
\alias{runUMAP}
\title{Cluster Data by UMAP}
\usage{
runUMAP(
  TapestriExperiment,
  alt.exp = "alleleFrequency",
  assay = NULL,
  use.pca.dims = TRUE,
  pca.dims = NULL,
  ...
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{alt.exp}{Chr string indicating altExp to use, NULL uses top-level experiment. Default "alleleFrequency".}

\item{assay}{Chr string indicating assay to use. NULL (default) selects first listed assay. Npt used when \code{use.pca.dims = TRUE}.}

\item{use.pca.dims}{Logical, if TRUE, uses experiment PCA, otherwise uses assay data. Default TRUE.}

\item{pca.dims}{Numeric vector, indicating indices of PCs to use in UMAP. NULL (default) uses all dimensions.}

\item{...}{Additional parameters to pass to umap, e.g. for configuration (see \code{\link[umap:umap.defaults]{umap::umap.defaults}}).}
}
\value{
TapestriExperiment with UMAP embeddings saved to reducedDims slot of altExp.
}
\description{
Cluster Data by UMAP
}
\examples{
\dontrun{TapestriExperiment <- runUMAP(TapestriExperiment, input.dims = 1:3)}
}
