% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyCalculation.R
\name{smoothPloidy}
\alias{smoothPloidy}
\title{Smooth ploidy values by chromosome and chromosome arm}
\usage{
smoothPloidy(TapestriExperiment, method = "median")
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object.}

\item{method}{Character string, smoothing method. Supports median (default) or mean.}
}
\value{
\code{TapestriExperiment} with \code{smoothPloidy} and \code{DiscretePloidy} assays in \code{altExp} slot
}
\description{
\code{smoothPloidy()} takes ploidy values across probes and smooths them by median (default) for each chromosome and chromosome arm,
resulting in one ploidy value per chromosome/arm per cell barcode.
Values are then discretized into integers by conventional rounding.
Results are stored in \code{altExp} slots in \code{TapestriObject}, with \code{smoothPloidy} and \code{DiscretePloidy} assays.
}
\examples{
\dontrun{
TapestriExperiment <- smoothPloidy(TapestriExperiment)
}
}
