% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSpecialtyProbes.R
\name{splitSpecialtyProbes}
\alias{splitSpecialtyProbes}
\title{Move specialty probe counts to alternative experiment slots}
\usage{
splitSpecialtyProbes(
  tapestri.experiment.object,
  separate.specialty.probes = c("grna", "sample.barcode", "Y")
)
}
\arguments{
\item{tapestri.experiment.object}{TapestriExperiment object}

\item{separate.specialty.probes}{Chr vector including a combination of "grna", "sample.barcode", and or "Y" (default), or FALSE.}
}
\value{
A \code{TapestriExperiment} with \code{altExp} slots filled with counts and metadata for the specified probes.
}
\description{
\code{splitSpecialtyProbes()} takes the probe IDs corresponding to \code{grnaProbe} and \code{barcodeProbe} slots of the \code{TapestriExperiment} object, as well as probes on chrY, and moves them to their own \code{altExp} (alternative experiment) slots in the object.
This allows those counts and associated metadata to be manipulated separately without interfering with the other probes in the panel.
}
\examples{
\dontrun{tapObject <- splitSpecialtyProbes(tapObject)}
}
\seealso{
\code{\link[SingleCellExperiment:splitAltExps]{SingleCellExperiment::splitAltExps()}} for manual specification of probes to move to \code{altExp} slots.
}
