% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestriExperimentDefinition.R
\docType{class}
\name{TapestriExperiment-class}
\alias{TapestriExperiment-class}
\alias{.TapestriExperiment}
\title{TapestriExperiment Class Definition}
\value{
TapestriExperiment
}
\description{
TapestriExperiment Class Definition
}
\section{Slots}{

\describe{
\item{\code{barcodeProbe}}{character.}

\item{\code{grnaProbe}}{character.}
}}

\examples{
x <- new("TapestriExperiment")
}
