% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assayBoxPlot}
\alias{assayBoxPlot}
\title{Generate a box plot from matrix data}
\usage{
assayBoxPlot(
  TapestriExperiment,
  alt.exp = NULL,
  assay = NULL,
  log.y = T,
  split.features = F,
  coldata.set = NULL
)
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object}

\item{alt.exp}{Character, \code{altExp} to plot. \code{NULL} (default) uses the top-level experiment in \code{TapestriExperiment}.}

\item{assay}{Character, assay to plot. \code{NULL} (default) selects first assay listed \code{TapestriExperiment}.}

\item{log.y}{Logical, if \code{TRUE}, scales data using \code{log1p()}. Default \code{TRUE.}}

\item{split.features}{Logical, if \code{TRUE}, splits plot by \code{rowData} features. Default \code{FALSE.}}

\item{coldata.set}{Character, \code{colData} field to use for X axis categories. Default \code{NULL}.}
}
\value{
A ggplot object using \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}.
}
\description{
Draws box plot of matrix data stored in a \code{TapestriExperiment}.
This is especially useful for visualizing \code{altExp} count data, such as counts from
probes on chrY.
}
\examples{
\dontrun{
assayBoxPlot(TapestriExperiment, "chrYCounts",
  assay = "counts",
  split.features = T, coldata.set = "cluster"
)
}
}
