% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assayHeatmap}
\alias{assayHeatmap}
\title{Generate Assay Heatmap}
\usage{
assayHeatmap(
  TapestriExperiment,
  alt.exp = NULL,
  assay = NULL,
  split.col.by = NULL,
  split.row.by = NULL,
  annotate.row.by = NULL
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{alt.exp}{Chr string indicating altExp slot to pull from. \code{NULL} (default) pulls from top-level/main experiment.}

\item{assay}{Chr string indicating assay slot to pull from. \code{NULL} (default) pulls from first-indexed assay (often "counts").}

\item{split.col.by}{Chr string indicating RowData field to split columns by, usually "chr" or "arm". Default NULL.}

\item{split.row.by}{Chr string indicating ColData field to split rows by, usually "cluster". Default NULL.}

\item{annotate.row.by}{Chr string indicating ColData field to use as annotation. Default NULL.}
}
\value{
A ComplexHeatmap object
}
\description{
Use to create a heatmap of matrix data in a TapestriObject.
Heatmaps are generated as transposed representations of the matrix.
}
\examples{
\dontrun{assayHeatmap(TapestriExperiment, alt.exp = "smoothedPloidyByArm",
assay = "discretePloidy", split.row.by = "cluster")}
}
