% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assayHeatmap}
\alias{assayHeatmap}
\title{Generate heatmap of matrix data}
\usage{
assayHeatmap(
  TapestriExperiment,
  alt.exp = NULL,
  assay = NULL,
  split.col.by = NULL,
  split.row.by = NULL,
  annotate.row.by = NULL,
  color.preset = NULL,
  color.custom = NULL,
  ...
)
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object}

\item{alt.exp}{Character, \code{altExp} slot to use. \code{NULL} (default) uses top-level/main experiment.}

\item{assay}{Character, \code{assay} slot to use. \code{NULL} (default) uses first-indexed assay (usually "counts").}

\item{split.col.by}{Character, \code{rowData} column to split columns by, usually "chr" or "arm". Default \code{NULL}.}

\item{split.row.by}{Character, \code{colData} column to split rows by, usually "cluster". Default \code{NULL}.}

\item{annotate.row.by}{Character, \code{colData} column to use for annotation. Default \code{NULL}.}

\item{color.preset}{Character, color preset to use to color heatmap, either "ploidy" or "ploidy.denoise" (see \code{Details}). Overrides \code{color.custom}. \code{NULL} (default) uses default \code{ComplexHeatmap} coloring.}

\item{color.custom}{Color mapping function given by \code{\link[circlize:colorRamp2]{circlize::colorRamp2()}}. \code{color.preset} must be \code{NULL}.}

\item{...}{Additional parameters to pass to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}
}
\value{
A \code{ComplexHeatmap} object
}
\description{
Creates a heatmap of matrix data in a \code{TapestriObject} using the \code{ComplexHeatmap} package.
Heatmaps are generated as transposed (i.e. x-y flipped) representations of the indicated matrix.
}
\section{\code{color.preset} Options}{
\subsection{"ploidy"}{

Blue-white-red gradient from 0-2-4. 4 to 8+ is red-black gradient.

\if{html}{\out{<div class="sourceCode">}}\preformatted{circlize::colorRamp2(c(0,1,2,3,4,8),
c('#2c7bb6','#abd9e9','#ffffff','#fdae61','#d7191c', "black"))
}\if{html}{\out{</div>}}
}

\subsection{"ploidy.denoise"}{

Similar to ploidy, but white range is from 1.5-2.5 to reduce the appearance of noise around diploid cells.

\if{html}{\out{<div class="sourceCode">}}\preformatted{circlize::colorRamp2(c(0,1,1.5,2,2.5,3,4,8),
c('#2c7bb6','#abd9e9','#ffffff','#ffffff','#ffffff','#fdae61','#d7191c', "black"))
}\if{html}{\out{</div>}}
}
}

\examples{
\dontrun{
assayHeatmap(TapestriExperiment,
  alt.exp = "smoothedCopyNumberByArm",
  assay = "discreteCopyNumber", split.row.by = "cluster"
)
}
}
\seealso{
\link[ComplexHeatmap]{Heatmap}
}
