% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calcNormCounts}
\alias{calcNormCounts}
\title{Normalize raw counts}
\usage{
calcNormCounts(TapestriExperiment, method = "mb")
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object.}

\item{method}{Character, normalization method. Default "mb".}
}
\value{
\code{TapestriExperiment} object with normalized counts added to \code{normcounts} slot.
}
\description{
Normalizes raw counts from \code{counts} slot in \code{TapestriExperiment} and returns the object with normalized counts in the \code{normcounts} slot.
"mb" is the only method supported currently.
This method performs the same normalization scheme as in Mission Bio's mosaic package for python.
Counts for each barcode are normalized relative to their barcode's mean and probe counts are normalized relative to their probe's median.
Also calculates the standard deviation for each probe using normalized counts and adds it to \code{rowData}.
}
\examples{
\dontrun{TapestriExperiment <- calcNormCounts(TapestriExperiment)}
}
