% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBarcodedReads.R
\name{callSampleLables}
\alias{callSampleLables}
\title{Call sample labels based on feature counts.}
\usage{
callSampleLables(
  TapestriExperiment,
  coldata.labels,
  sample.label = "sample.call",
  return.table = F,
  neg.label = NA,
  method = "max",
  ties.method = "first"
)
}
\arguments{
\item{TapestriExperiment}{A \code{TapestriExperiment} object}

\item{coldata.labels}{Character vector, column fields corresponding to \code{colData}.}

\item{sample.label}{Character, column name to use for the sample call. Default "sample.call".}

\item{return.table}{Logical, if \code{TRUE}, returns a data.frame of the sample.calls. If \code{FALSE} (default), returns updated \code{TapestriExperiment} object.}

\item{neg.label}{Character, label for samples with no counts. Default \code{NA}.}

\item{method}{Character, call method. Only "max" currently supported, calls based on whichever \code{coldata.labels} field has the most counts.}

\item{ties.method}{Character, passed to \code{max.col()} indicating how to break ties. Default "first".}
}
\value{
A \code{TapestriExperiment} object with sample calls added to \code{colData.} If \code{return.table == TRUE}, a data.frame of sample calls.
}
\description{
\code{callSampleLables()} determines sample labels by comparing auxiliary feature count data,
most likely generated from barcoded reads (see \link{parseBarcodedReads}).
For \code{method = max}, labels are dictated by whichever \code{coldata.labels} field has the highest number of counts.
By default, ties are broken by choosing whichever label has the lowest index position (\code{ties.method = "first"}).
Samples with 0 counts for all \code{coldata.labels} fields are labeled according to \code{neg.label}.
}
\examples{
\dontrun{
TapestriExperiment <- callSampleLables(TapestriExperiment,
  coldata.labels = c("g7", "gNC"),
  sample.label = "sample.grna"
)
}
}
