% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNweaveR-package.R
\docType{package}
\name{CNweaveR-package}
\alias{CNweaveR}
\alias{CNweaveR-package}
\title{CNweaveR: DNA CNV Analysis for Custom Genome-Wide Tapestri Panels}
\description{
This package enables basic analysis of CNVs using custom genome-wide targeted DNA sequencing panels for the Mission Bio Tapestri system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joeymays/weaver}
  \item Report bugs at \url{https://github.com/joeymays/weaver/issues}
}

}
\author{
\strong{Maintainer}: Joseph Mays \email{josephcmays@gmail.com} (\href{https://orcid.org/0000-0003-4903-938X}{ORCID})

}
\keyword{internal}
