% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{corner}
\alias{corner}
\title{Print the top-left corner of a matrix}
\usage{
corner(input.mat)
}
\arguments{
\item{input.mat}{A matrix-like object.}
}
\value{
A matrix-like object matching input class, subset to a maximum of 5 rows and columns.
}
\description{
Outputs up to 5 rows and columns of the input matrix object (with \code{rownames} and \code{colnames}) to get a quick look without filling the console.
}
\examples{
\dontrun{
corner(assay(TapestriObject))
}
}
