% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTapestriExperiment.R
\name{createTapestriExperiment}
\alias{createTapestriExperiment}
\title{Create TapestriExperiment object from Tapestri Pipeline output}
\usage{
createTapestriExperiment(h5.filename, panel.id = "")
}
\arguments{
\item{h5.filename}{file path for .h5 file form Tapestri Pipeline output}

\item{panel.id}{Tapestri panel name, CO261 and CO293 supported only.}
}
\value{
TapestriExperiment
}
\description{
Create TapestriExperiment object from Tapestri Pipeline output
}
\examples{
\dontrun{x <- createTapestriExperiment("myh5file.h5", "CO293")}
}
