% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTapestriExperiment.R
\name{createTapestriExperiment}
\alias{createTapestriExperiment}
\title{Create TapestriExperiment object from Tapestri Pipeline output}
\usage{
createTapestriExperiment(
  h5.filename,
  panel.id = NULL,
  get.cytobands = TRUE,
  genome = "hg19",
  move.non.genome.probes = c("grna", "sample.barcode", "Y"),
  filter.variants = T
)
}
\arguments{
\item{h5.filename}{file path for .h5 file from Tapestri Pipeline output.}

\item{panel.id}{Tapestri panel name, CO261 and CO293 supported only. Default NULL.}

\item{get.cytobands}{Logical value indicating whether to retrieve and add chromosome cytobands and chromosome arms to probe metadata. Default TRUE.}

\item{genome}{Chr string indicating reference genome to pull cytobands and arms from. Only hg19 is currently supported. Default "hg19".}

\item{move.non.genome.probes}{Chr vector indicating non-genomic probes to move counts and metadata to altExp slots. Default c("grna", "sample.barcode", "Y").}

\item{filter.variants}{Logical. If TRUE, only loads variants that have passed Tapestri Pipeline filters. Default TRUE.}
}
\value{
Constructed TapestriExperiment object
}
\description{
\code{createTapestriExperiment()} constructs a \code{TapestriExperiment} object from \code{.h5} file output by the Tapestri pipeline.
Probe metadata is automatically imported by default.
\code{panel.id} is an optional shortcut to set gRNA and barcode probe identities.
\code{move.non.genome.probes} moves the specified probes to altExp (alternative experiment) slots in the TapestriExperiment object.
Probes corresponding to the \code{barcodeProbe} and \code{grnaProbe} slots, and probes on ChrY are moved by default.
Basic QC stats (e.g. total number of reads per probe) are added to either the object's colData or rowData.
Basic metadata is automatically to the \code{metadata} slot.
\code{filter.variants == TRUE} only loads variants that have passed filters in the Tapestri Pipeline.
This greatly reduces the number of variants to a smaller number of useful ones.
}
\examples{
\dontrun{x <- createTapestriExperiment("myh5file.h5", "CO293")}
}
\seealso{
\code{\link[=moveNonGenomeProbes]{moveNonGenomeProbes()}}, \code{\link[=getCytobands]{getCytobands()}} which are run as part of this function for convenience.
}
