% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTapestriExperiment.R
\name{CreateTapestriExperiment}
\alias{CreateTapestriExperiment}
\title{Create TapestriExperiment object from Tapestri Pipeline output}
\usage{
CreateTapestriExperiment(h5.filename, panel.id = "")
}
\description{
Create TapestriExperiment object from Tapestri Pipeline output
}
