% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTapestriExperiment.R
\name{createTapestriExperiment}
\alias{createTapestriExperiment}
\title{Create TapestriExperiment object from Tapestri Pipeline output}
\usage{
createTapestriExperiment(
  h5.filename,
  panel.id = NULL,
  get.cytobands = TRUE,
  genome = "hg19",
  separate.specialty.probes = c("grna", "sample.barcode", "Y")
)
}
\arguments{
\item{h5.filename}{file path for .h5 file from Tapestri Pipeline output.}

\item{panel.id}{Tapestri panel name, CO261 and CO293 supported only. Default NULL.}

\item{get.cytobands}{Logical value indicating whether to retrieve and add chromosome cytobands and chromosome arms to probe metadata.}

\item{genome}{Chr string indicating reference genome to pull cytobands and arms from. Only hg19 is currently supported.}

\item{separate.specialty.probes}{Chr vector indicating specialty probes to move to altExp slots.}
}
\value{
Constructed TapestriExperiment object
}
\description{
\code{createTapestriExperiment()} constructs a \code{TapestriExperiment} object from \code{.h5} file output by the Tapestri pipeline.
Probe metadata is automatically imported by default.
\code{panel.id} is an optional shortcut to set gRNA and barcode probe identities.
\code{separate.specialty.probes} moves the specified probes to altExp (alternative experiment) slots in the TapestriExperiment object.
Probes corresponding to the \code{barcodeProbe} and \code{grnaProbe} slots, and probes on ChrY are moved by default.
}
\examples{
\dontrun{x <- createTapestriExperiment("myh5file.h5", "CO293")}
}
\seealso{
\code{\link[=splitSpecialtyProbes]{splitSpecialtyProbes()}}, \code{\link[=getCytobands]{getCytobands()}} which are run as part of this function for convenience.
}
