% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getClusters}
\alias{getClusters}
\title{Cluster data}
\usage{
getClusters(
  TapestriExperiment,
  feature.set = "alleleFrequency",
  dim.reduction = "UMAP",
  eps = 0.8,
  ...
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{feature.set}{Chr string identifying the altExp feature set to use Default "alleleFrequency".}

\item{dim.reduction}{Chr string indicating the reduced dimension set to use. Default "UMAP".}

\item{eps}{Numeric, dbscan eps parameter. Change to adjust cluster granularity. See \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Default 0.8.}

\item{...}{Additional parameters to pass to \code{dbscan::dbscan()}.}
}
\value{
TapestriExperiment object with updated colData
}
\description{
Cluster data using dbscan method. Uses UMAP reduced dimensions to partition data into clusters and saves the clusters to colData.
}
\examples{
\dontrun{TapestriExperiment <- getClusters(TapestriExperiment, dim.reduction = "UMAP", eps = 0.8)}
}
\seealso{
\code{\link[dbscan:dbscan]{dbscan::dbscan()}}.
}
