% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtureCalls.R
\name{getGMMBoundaries}
\alias{getGMMBoundaries}
\title{getGMMBoundaries}
\usage{
getGMMBoundaries(TapestriExperiment, chromosome.scope = "chr")
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object}

\item{chromosome.scope}{}
}
\description{
getGMMBoundaries
}
