% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyCalculation.R
\name{getPloidy}
\alias{getPloidy}
\alias{generateControlPloidyTemplate}
\title{Calculate cell-chromosome ploidy using control cell population as baseline}
\usage{
getPloidy(TapestriExperiment, control.ploidy, coldata.set = "cluster")

generateControlPloidyTemplate(ploidy.all = 2, sample.label.all = "cluster")
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object.}

\item{control.ploidy}{A \code{data.frame} with columns \code{arm}, \code{ploidy}, and \code{sample.label}. See description.}

\item{coldata.set}{Character, \code{colData} field to use for subsetting cell.barcodes. Default "cluster".}

\item{ploidy.all}{Numeric, sets all entries of \code{ploidy} column in output. Default 2.}

\item{sample.label.all}{Character, sets all entries of \code{sample.label} column in output. Default "cluster".}
}
\value{
\code{TapestriExperiment} object with ploidy values in \code{ploidy} assay slot.

\code{data.frame} with 3 columns named \code{arm}, \code{ploidy}, and \code{sample.label}
}
\description{
\code{getPloidy()} transforms the normalized count matrix of a \code{TapestriExperiment} object
into ploidy values based on a reference subset of cell barcodes and given ploidy value (e.g. 2 for diploid).
This is practically used to set the median ploidy of a control, usually diploid,
cell population to a known ploidy value, e.g. 2, and then calculate the ploidy for all the
cells relative to that control population. This occurs individually for each probe.
\code{control.ploidy} is a \code{data.frame} lookup table used to indicate the ploidy value and cell barcodes
to use as the reference. A template for \code{control.ploidy} can be generated using \code{\link[=generateControlPloidyTemplate]{generateControlPloidyTemplate()}}.
The \code{control.ploidy} data.frame should include 3 columns named \code{arm}, \code{ploidy}, and \code{sample.label}.
\code{arm} is chromosome arms from chr1p through chrXq, \code{ploidy} is the reference ploidy value, and \code{sample.label} is the
value corresponding to the \code{colData} field given in \code{coldata.set} to indicate the cell barcode subset to use to set the ploidy.
}
\section{Functions}{
\itemize{
\item \code{generateControlPloidyTemplate()}: generates a \code{data.frame} template for use in \code{getPloidy()}.

}}
\examples{
\dontrun{
TapestriExperiment <- getPloidy(TapestriExperiment,
  control.ploidy,
  coldata.set = "cluster"
)
}
\dontrun{
control.ploidy <- generateControlPloidyTemplate()
}
}
