% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getTidyData}
\alias{getTidyData}
\title{Get tidy-style data from TapestriExperiment objects}
\usage{
getTidyData(TapestriExperiment, alt.exp = NULL, assay = NULL)
}
\arguments{
\item{TapestriExperiment}{A TapestriExperiment object}

\item{alt.exp}{Chr string indicating altExp slot to pull from. \code{NULL} (default) pulls from top-level/main experiment.}

\item{assay}{Chr string indicating assay slot to pull from. \code{NULL} (default) pulls from first-indexed assay (often "counts").}
}
\value{
A tibble of tidy data with corresponding metadata from colData and rowData.
}
\description{
\code{getTidyData()} pulls the matrix from the indicated assay and/or altExp slot(s), and morphs it into tidy format.
ColData from the top-level/main experiment is merged.
RowData from the indicated assay and/or altExp slot(s) is merged.
}
\examples{
\dontrun{getTidyData(TapestriObject, alt.exp = "alleleFrequency")}
}
