% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBarcodedReads.R
\name{parseBarcodedReadsFromContig}
\alias{parseBarcodedReadsFromContig}
\alias{parseBarcodedReads}
\title{Parse Barcoded Reads}
\usage{
parseBarcodedReadsFromContig(
  bam.file,
  barcode.lookup,
  contig,
  cell.barcode.tag = "RG",
  max.mismatch = 2,
  with.indels = T
)

parseBarcodedReads(
  tapestri.experiment.object,
  bam.file,
  barcode.lookup,
  probe,
  ...
)
}
\arguments{
\item{bam.file}{Chr string indicating file path of BAM file. \code{.bai} BAM index file must be in the same location.}

\item{barcode.lookup}{data.frame where the first column is the barcode identifier/name and the second column is the DNA sequence. Headers are ignored.}

\item{contig}{Chr string of contig or chromosome name for search.}

\item{cell.barcode.tag}{Character string of length 2, indicates cell barcode field in BAM, specified by Tapestri pipeline (currently "RG"). Default "RG".}

\item{max.mismatch}{The maximum and minimum number of mismatching letters allowed. See \code{\link[Biostrings:matchPattern]{Biostrings::matchPattern()}}.}

\item{with.indels}{If TRUE then indels are allowed. See \code{\link[Biostrings:matchPattern]{Biostrings::matchPattern()}}.}

\item{tapestri.experiment.object}{TapestriExperiment}

\item{probe}{Chr string, either "gRNA" or "sample.barcode" to parse counts from grnaCounts or sampleBarcodeCounts alternative experiments, respectively.}

\item{...}{Arguments to pass on to \code{parseBarcodedReadsFromContig()}.}
}
\value{
A data.frame of read counts for each specified barcode.

A data.frame of read counts for each specified barcode
}
\description{
\code{parseBarcodedReads()} and \code{parseBarcodedReadsFromContig()} match exogenous DNA barcode sequences from plasmid transduction to their associated cell barcodes and outputs them as a table of counts.
\code{parseBarcodedReads()} is a shortcut for \code{parseBarcodedReadsFromContig()}, allowing the user to specify 'gRNA' or 'sample.barcode'.
}
\examples{
\dontrun{counts <- parseBarcodedReadsFromContig(bam.file, barcode.lookup, "virus_ref2")}
\dontrun{counts <- parseBarcodedReads(tapestri.experiment.object,
bam.file, barcode.lookup, "gRNA")}
}
