% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{PCAKneePlot}
\alias{PCAKneePlot}
\title{Plot PCA Variance Knee}
\usage{
PCAKneePlot(TapestriExperiment, alt.exp = "alleleFrequency", n.pcs = 10)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{alt.exp}{Chr string indicating altExp to use, NULL uses top-level experiment. Default "alleleFrequency".}

\item{n.pcs}{Numeric vector, indicating number of PCs to plot, starting at 1. Default 10.}
}
\value{
ggplot knee plot
}
\description{
Draws knee plot of PCA variance explained to determine which PCs to include for downstream applications e.g. clustering.
}
\examples{
\dontrun{PCAKneePlot(TapestriExperiment, pcs = 1:5)}
}
