% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtureCalls.R
\name{plotCopyNumberGMM}
\alias{plotCopyNumberGMM}
\title{Plot copy number GMM components}
\usage{
plotCopyNumberGMM(
  TapestriExperiment,
  feature.id = 1,
  chromosome.scope = "chr",
  draw.boundaries = F
)
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object}

\item{feature.id}{chromosome or chromosome arm}

\item{chromosome.scope}{"chr" or "arm", for plotting models for either whole chromosomes or chromosome arms}

\item{draw.boundaries}{logical, if \code{TRUE}, draw decision boundaries between each Gaussian component}
}
\value{
\code{ggplot} object
}
\description{
Plot copy number GMM components
}
\examples{
\dontrun{
TapestriExperiment <- plotCopyNumberGMM(TapestriExperiment,
  feature.id = 7,
  chromosome.scope = "chr",
  draw.boundaries = T
)
}
}
