% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reducedDimPlot}
\alias{reducedDimPlot}
\title{Scatter plot for reduced dimensions}
\usage{
reducedDimPlot(
  TapestriExperiment,
  plot.reduction,
  pc.x = 1,
  pc.y = 2,
  feature.set = "alleleFrequency",
  group.label = NULL
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{plot.reduction}{Chr, which dimension reduction to plot, either "PCA" or "UMAP".}

\item{pc.x}{Numeric, if \code{plot.reduction == "PCA"}, index of PC to plot. Default 1 for PC1.}

\item{pc.y}{Numeric, if \code{plot.reduction == "PCA"}, index of PC to plot. Default 2 for PC2.}

\item{feature.set}{Chr string identifying the altExp feature set to plot. Default "alleleFrequency".}

\item{group.label}{Chr string indicating colData column for coloring samples. Default NULL.}
}
\value{
ggplot scatter plot
}
\description{
Scatter plot for reduced dimensions
}
\examples{
\dontrun{reducedDimPlot(TapestriExperiment, plot.reduction = "pca")}
}
