% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{runUMAP}
\alias{runUMAP}
\title{Cluster Data by UMAP}
\usage{
runUMAP(
  TapestriExperiment,
  feature.set = "alleleFrequency",
  use.pca.dims = T,
  input.dims = NULL,
  ...
)
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object}

\item{feature.set}{Chr string identifying the altExp feature set to perform UMAP on. Default "alleleFrequency".}

\item{use.pca.dims}{Logical, only TRUE supported. Default TRUE.}

\item{input.dims}{Numeric vector, indicating indices of PCs to use in UMAP.}

\item{...}{Additional parameters to pass to umap, e.g. for configuration (see \code{\link[umap:umap.defaults]{umap::umap.defaults}}).}
}
\value{
TapestriExperiment with UMAP embeddings saved to reducedDims slot of altExp.
}
\description{
Cluster Data by UMAP
}
\examples{
\dontrun{TapestriExperiment <- runUMAP(TapestriExperiment, input.dims = 1:3)}
}
