% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestriExperimentDefinition.R
\name{barcodeProbe}
\alias{barcodeProbe}
\alias{barcodeProbe,TapestriExperiment-method}
\alias{barcodeProbe<-}
\alias{barcodeProbe<-,TapestriExperiment-method}
\alias{grnaProbe}
\alias{grnaProbe,TapestriExperiment-method}
\alias{grnaProbe<-}
\alias{grnaProbe<-,TapestriExperiment-method}
\title{Getters and Setters for TapestriExperiment slots}
\usage{
barcodeProbe(x)

\S4method{barcodeProbe}{TapestriExperiment}(x)

barcodeProbe(x) <- value

\S4method{barcodeProbe}{TapestriExperiment}(x) <- value

grnaProbe(x)

\S4method{grnaProbe}{TapestriExperiment}(x)

grnaProbe(x) <- value

\S4method{grnaProbe}{TapestriExperiment}(x) <- value
}
\arguments{
\item{x}{A TapestriExperiment object}

\item{value}{Chr string to assign to slot}

\item{TapestriExperiment}{A TapestriExperiment object}
}
\description{
Current slots supported are barcodeProbe for a sample barcode probe ID and grnaProbe for a gRNA-associated probe ID.
}
\section{Functions}{
\itemize{
\item \code{barcodeProbe()}: Getter and Setter functions for custom slots in TapestriExperiment Object.

}}
