% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyCalculation.R
\name{smoothPloidy}
\alias{smoothPloidy}
\title{Smooth Ploidy Values by Chromosome and Chromosome Arm}
\usage{
smoothPloidy(TapestriExperiment, method = "median")
}
\arguments{
\item{TapestriExperiment}{TapestriExperiment object.}

\item{method}{Chr string indicating smoothing method. Supports median (default) or mean.}
}
\value{
TapestriExperiment object with smoothed ploidy values in altExp slot
}
\description{
\code{smoothPloidy()} takes ploidy values across probes and smooths them by median (default) for each chromosome and chromosome arm, resulting in one ploidy value per chromosome/arm per cell barcode.
Values are then discretized into integers by conventional rounding.
Results are stored as alternate experiments in the TapestriObject, with smoothPloidy and DiscretePloidy assays.
}
\examples{
\dontrun{TapestriExperiment <- smoothPloidy(TapestriExperiment)}
}
