% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidyCalculation.R
\name{smoothPloidy}
\alias{smoothPloidy}
\title{Smooth ploidy values by chromosome and chromosome arm}
\usage{
smoothPloidy(TapestriExperiment, method = "median")
}
\arguments{
\item{TapestriExperiment}{\code{TapestriExperiment} object.}

\item{method}{Character, smoothing method. Supports median (default) or mean.}
}
\value{
\code{TapestriExperiment} with \code{smoothPloidy} and \code{DiscretePloidy} assays in \code{altExp} slots \code{smoothedPloidyByChr} and \code{smoothedPloidyByArm}.
}
\description{
\code{smoothPloidy()} takes \code{ploidy} slot values for probes on a chromosome and smoothes them by median (default) for each chromosome
and chromosome arm, resulting in one ploidy value per chromosome and chromosome arm for each cell barcode.
Values are then discretized into integers by conventional rounding.
Smoothed ploidy and discretized smoothed ploidy values are stored as \code{smoothPloidy} and \code{DiscretePloidy} assays,
in \code{altExp} slots \code{smoothedPloidyByChr} for chromosome-level smoothing, and \code{smoothedPloidyByArm} for chromosome arm-level smoothing.
}
\examples{
\dontrun{
TapestriExperiment <- smoothPloidy(TapestriExperiment)
}
}
